Added alternate way to instantiate validator
The current code requires you to pass a file path when constructing
the Validator; however, sometimes you don't have the schema in a file
(e.g. it's in a database). This offers a second way to instantiate
it, by passing in the raw JSON schema, and specifying it as "json"
instead of a file location.